Incorporating review comments
