+++ bundles/af-core/src/main/java/com/adobe/cq/forms/core/components/internal/models/v2/form/FormContainerImpl.java
@@ -18,6 +18,8 @@ package com.adobe.cq.forms.core.components.internal.models.v2.form;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import javax.annotation.PostConstruct;
+
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.resource.Resource;
@@ -76,6 +78,8 @@ public class FormContainerImpl extends AbstractContainerImpl implements
     @Nullable
     private String clientLibRef;
 
+    protected String contextPath = StringUtils.EMPTY;
+
     @ValueMapValue(injectionStrategy = InjectionStrategy.OPTIONAL)
     @Nullable
     private String title;
@@ -91,6 +95,23 @@ public class FormContainerImpl extends AbstractContainerImpl implements
     @ValueMapValue(injectionStrategy = InjectionStrategy.OPTIONAL)
     @Nullable
     private String data;
+
+    @PostConstruct
+    protected void initFormContainerModel() {
+        if (request != null) {
+            contextPath = request.getContextPath();
+        }
+    }
+
+    @Override
+    public void setContextPath(String contextPath) {
+        this.contextPath = contextPath;
+    }
+
+    @JsonIgnore
+    public String getContextPath() {
+        return contextPath;
+    }
 
     @Override
     @Nullable
